+++ packages/stream_video_flutter/lib/src/call_screen/call_content/picture_in_picture/android_pip_overlay.dart
@@ -55,10 +55,9 @@ class _AndroidPipOverlayState extends State<AndroidPipOverlay>
     super.initState();
     recalculateParticipants(widget.call.state.value.callParticipants);
 
-    _participantsSubscription = widget.call
-        .partialState((state) => state.callParticipants)
-        .throttleByCollectionSize()
-        .listen(recalculateParticipants);
+    _participantsSubscription = widget.call.participantsStream.listen(
+      recalculateParticipants,
+    );
   }
 
   @override